refactor(transport): split transport.py into ref / transport / worker_local
937-line transport.py conflated the data model with the backend machinery. Split:
- ref.py: TensorHandle / TensorSpan / TensorRef / cat_rows / map_tree (pure data
  model + addressing, Ray-free; transform/materialize reach the runtime lazily)
- transport.py: TensorTransport ABC + dehydrate/hydrate + TransportSession +
  TensorTransportRuntime + _collect_leaves
- worker_local.py: WorkerLocalTransport + the localize find/move/replace NCCL
  routing + _apply_tensor_op
Importers repointed: inside-tensor modules use submodule imports (acyclic),
outside modules + tests use the package facade. Pure move; no behavior change.
