+++ tests/transport/test_tree_walk.py
@@ -26,13 +26,8 @@ import pytest
 import torch
 
 from unirl.distributed.tensor.batch import Batch, packed_field, shared_field
-from unirl.distributed.tensor.transport import (
-    TensorRef,
-    TensorSpan,
-    TensorTransport,
-    _collect_leaves,
-    map_tree,
-)
+from unirl.distributed.tensor import TensorRef, TensorSpan, TensorTransport, map_tree
+from unirl.distributed.tensor.transport import _collect_leaves
 
 pytestmark = pytest.mark.cpu
