+++ unirl/distributed/tensor/transport.py
@@ -1,24 +1,13 @@
-"""TensorTransport: backend-agnostic tensor storage and retrieval.
+"""TensorTransport — the backend-agnostic storage/retrieval contract.
 
-``TensorRef`` is the universal tensor proxy — a ``Batch`` subclass that
-holds an ordered list of :class:`TensorSpan`, each a contiguous row-window
-over one backend handle (from a single ``put()`` call). ``spans`` is the
-concat axis; per-span row counts derive ``sizes`` and ``batch_size`` (no
-stored size array). Per-row ``select`` / ``slice`` builds new spans — no data
-motion, no hydration. A :class:`TensorSpan` is generic over its handle type,
-bound by the :class:`TensorHandle` protocol.
-
-``TensorRef`` also serves as a compute proxy: ``transform``, ``reshape``,
-``permute``, ``local`` delegate to the active backend.
-
-``TensorTransport`` is the ABC every backend implements: per-tensor ``put`` /
-``get``, optional ``put_batch`` / ``get_batch`` overrides, and a generic
-``transform`` for remote compute. It also owns the tree-walking
-``dehydrate`` / ``hydrate`` methods and a ``session`` context manager for
-cross-object batching.
-
-``TensorTransportRuntime`` is the per-process singleton that call sites
-reach for at runtime.
+The ABC every backend implements: per-tensor ``put`` / ``get`` (+ optional
+batched overrides), the tree-walking ``dehydrate`` / ``hydrate`` boundary, a
+``session`` context for cross-object batching, and a generic ``transform`` for
+remote compute. Worker-resident backends add their machinery via
+``unirl.distributed.tensor.worker_local.WorkerLocalTransport``;
+``TensorTransportRuntime`` is the per-process singleton call sites reach at
+runtime. The ``TensorRef`` / ``TensorSpan`` data model lives in
+``unirl.distributed.tensor.ref``.
 """
 
 from __future__ import annotations
@@ -26,332 +15,15 @@ from __future__ import annotations
 import abc
 import logging
 from contextlib import contextmanager
-from dataclasses import dataclass
 from dataclasses import fields as dc_fields
-from typing import (
-    Any,
-    Callable,
-    ClassVar,
-    Dict,
-    Generic,
-    Iterator,
-    List,
-    Optional,
-    Protocol,
-    Set,
-    Tuple,
-    TypeVar,
-    runtime_checkable,
-)
+from typing import Any, Callable, Dict, Iterator, List, Optional, Set, Tuple
 
-import ray
 import torch
 
-from unirl.distributed.tensor.batch import Batch, concat_field, shared_field
-from unirl.distributed.utils import collect_leaves
+from unirl.distributed.tensor.batch import Batch
+from unirl.distributed.tensor.ref import TensorRef, TensorSpan, cat_rows
 
 logger = logging.getLogger(__name__)
-
-
-@runtime_checkable
-class TensorHandle(Protocol):
-    """The minimal handle contract a :class:`TensorSpan` resolves through.
-
-    The worker-local store handle (``GPUTensorHandle`` / ``ColocateTensorHandle``)
-    and the global ``TQTensorHandle`` both satisfy it structurally. The
-    worker-local-only surface (``store_key`` / ``source_id`` / ``object_ref``) is
-    reached via ``span.handle`` in the backends that need it, never off the span —
-    so it is intentionally absent from this protocol.
-    """
-
-    def local(self) -> torch.Tensor: ...
-
-
-T = TypeVar("T", bound=TensorHandle)
-
-
-class TensorSpan(Generic[T]):
-    """A contiguous half-open ``[start:stop)`` row-window over one handle — the addressing primitive.
-
-    ``TensorRef`` selection/permutation emits these instead of moving data:
-    a span addresses ``handle[start:stop)`` along dim 0 while the bytes stay in
-    the producing worker's store. Backends resolve a span by resolving the
-    handle and slicing (zero-copy on the IPC/store path); ``localize`` ships
-    only the ``[start:stop)`` rows cross-device, not the whole handle block.
-
-    Lifecycle: a span holds a PYTHON REFERENCE to the handle object, so
-    CPython's own reference counting aggregates all spans' lifetimes onto the
-    handle — the handle's single GC finalizer fires only when the last span
-    (and the handle itself) is gone. Spans carry no finalizer, no store_key
-    of their own, and trigger no extra decref RPCs.
-
-    Nested spans flatten at construction (``TensorSpan(TensorSpan(h,2,8),1,3)``
-    is ``TensorSpan(h,3,5)``), so ``handle`` is always a backend handle.
-    """
-
-    __slots__ = ("handle", "start", "stop")
-    handle: T
-
-    def __init__(self, handle: T | "TensorSpan[T]", start: int, stop: int) -> None:
-        start, stop = int(start), int(stop)
-        if isinstance(handle, TensorSpan):
-            start, stop = handle.start + start, handle.start + stop
-            handle = handle.handle
-        if not (0 <= start <= stop):
-            raise ValueError(f"TensorSpan range [{start}, {stop}) is invalid")
-        self.handle = handle
-        self.start = start
-        self.stop = stop
-
-    @property
-    def nrows(self) -> int:
-        return self.stop - self.start
-
-    def __len__(self) -> int:
-        return self.stop - self.start
-
-    # ── delegated metadata: shape (sliced) / dtype / device — read by nccl_recv
-    #    and repr. Handle-specific identity (store_key/source_id/object_ref) is
-    #    reached via ``span.handle`` in the worker-local backends, not off the span.
-
-    @property
-    def shape(self) -> tuple:
-        handle_shape = tuple(self.handle.shape)
-        return (self.stop - self.start, *handle_shape[1:])
-
-    @property
-    def dtype(self):
-        return self.handle.dtype
-
-    @property
-    def device(self):
-        return self.handle.device
-
-    def local(self) -> torch.Tensor:
-        return self.handle.local()[self.start : self.stop]
-
-    def __getstate__(self) -> dict:
-        return {"handle": self.handle, "start": self.start, "stop": self.stop}
-
-    def __setstate__(self, state: dict) -> None:
-        self.handle = state["handle"]
-        self.start = state["start"]
-        self.stop = state["stop"]
-
-    def __repr__(self) -> str:
-        return f"TensorSpan({self.handle!r}[{self.start}:{self.stop}])"
-
-
-def cat_rows(parts: List[torch.Tensor]) -> torch.Tensor:
-    """Concatenate per-ref tensors along dim 0 — the single assembly funnel.
-
-    Trailing-dim CONTRACT: parts may be padded to different widths per
-    producing shard (e.g. per-worker prompt blocks); 2D+ parts are
-    right-padded with zeros to the max width before the cat — consumers of
-    2D+ per-shard-padded fields must be mask-driven (the convention
-    ``TextTokenCondition.concat`` already establishes).
-    """
-    if not parts:
-        return torch.empty(0)
-    if len(parts) == 1:
-        return parts[0]
-    if parts[0].dim() >= 2:
-        widths = {int(t.shape[1]) for t in parts}
-        if len(widths) > 1:
-            target = max(widths)
-            padded = []
-            for t in parts:
-                if int(t.shape[1]) < target:
-                    pad = t.new_zeros((t.shape[0], target - t.shape[1]) + tuple(t.shape[2:]))
-                    t = torch.cat([t, pad], dim=1)
-                padded.append(t)
-            parts = padded
-    return torch.cat(parts, dim=0)
-
-
-@dataclass
-class TensorRef(Batch):
-    """The dehydrated-tensor proxy — an ordered list of row-window spans.
-
-    Each element of ``spans`` is a :class:`TensorSpan` over one backend handle;
-    ``spans`` is the concat axis. Per-span row counts derive ``sizes`` and
-    ``batch_size`` — there is no stored size array. ``batch_size`` is the total
-    row count, not ``len(spans)``.
-    """
-
-    spans: List[TensorSpan] = concat_field(default_factory=list)
-    shape: Optional[Tuple[int, ...]] = shared_field(default=None)
-    dtype: Optional[torch.dtype] = shared_field(default=None)
-    device: Optional[str] = shared_field(default=None)
-    grad: Optional["TensorRef"] = shared_field(default=None)
-    retain_grad_flag: bool = shared_field(default=False)
-
-    @property
-    def sizes(self) -> List[int]:
-        return [s.stop - s.start for s in self.spans]
-
-    @property
-    def batch_size(self) -> int:
-        return sum(s.stop - s.start for s in self.spans)
-
-    @classmethod
-    def concat(cls, items: "list[TensorRef]") -> "TensorRef":
-        spans: List[TensorSpan] = []
-        for item in items:
-            spans.extend(item.spans)
-        return items[0]._from_spans(spans)
-
-    def select(self, indices) -> "TensorRef":
-        """Re-index along the row axis by building ref VIEWS (no data motion)."""
-        idx = [int(i) for i in (indices.tolist() if hasattr(indices, "tolist") else indices)]
-        return self._gather(idx)
-
-    def _offsets(self) -> List[int]:
-        offsets = [0]
-        for s in self.spans:
-            offsets.append(offsets[-1] + (s.stop - s.start))
-        return offsets
-
-    def select_ranges(self, ranges: List[Tuple[int, int]]) -> "TensorRef":
-        """Re-index by global ``[start, stop)`` row ranges — the single selection engine.
-
-        Splits each range at span boundaries into views: a boundary-aligned piece
-        reuses its span untouched (zero-cost); a partial piece wraps in a new
-        :class:`TensorSpan` (nested spans flatten in the ctor). No sort/dedup, so
-        out-of-order, overlapping, and empty range lists behave as given.
-        """
-        span_offsets = self._offsets()  # cumulative row boundaries; span_offsets[-1] == total rows
-        selected: List[TensorSpan] = []
-        for range_start, range_stop in ranges:
-            range_start, range_stop = int(range_start), int(range_stop)
-            if not (0 <= range_start <= range_stop <= span_offsets[-1]):
-                raise IndexError(
-                    f"row range [{range_start}, {range_stop}) out of bounds for size {span_offsets[-1]}"
-                )
-            cursor = range_start
-            span_idx = 0
-            while cursor < range_stop:
-                while span_offsets[span_idx + 1] <= cursor:
-                    span_idx += 1
-                take = min(range_stop, span_offsets[span_idx + 1]) - cursor
-                src_span = self.spans[span_idx]
-                local_start = cursor - span_offsets[span_idx]
-                local_stop = local_start + take
-                selected.append(
-                    src_span
-                    if (local_start == 0 and local_stop == src_span.nrows)
-                    else TensorSpan(src_span, local_start, local_stop)
-                )
-                cursor += take
-        return self._from_spans(selected)
-
-    def _from_spans(self, spans: List[TensorSpan]) -> "TensorRef":
-        """Build a ref over the given spans, deriving shape from the row total.
-
-        Deliberately does NOT carry ``grad`` / ``retain_grad_flag`` — matches the
-        existing constructor default for a freshly selected/sliced view.
-        """
-        total_rows = sum(span.nrows for span in spans)
-        return TensorRef(
-            spans=spans,
-            shape=(total_rows, *self.shape[1:]) if self.shape else None,
-            dtype=self.dtype,
-            device=self.device,
-        )
-
-    def _gather(self, idx: List[int]) -> "TensorRef":
-        """Arbitrary re-index (gather/permute): coalesce consecutive runs into ranges."""
-        ranges: List[Tuple[int, int]] = []
-        run_start = 0
-        while run_start < len(idx):
-            run_end = run_start + 1
-            while run_end < len(idx) and idx[run_end] == idx[run_end - 1] + 1:
-                run_end += 1
-            ranges.append((int(idx[run_start]), int(idx[run_end - 1]) + 1))
-            run_start = run_end
-        return self.select_ranges(ranges)
-
-    def with_spans(self, spans: List[Any]) -> "TensorRef":
-        """Clone with substituted (routed) spans, preserving shape/dtype/device.
-
-        ``localize`` rebuilds refs after routing; this keeps the substitution
-        structural (same span count, derived sizes, no re-derivation).
-        """
-        return TensorRef(
-            spans=list(spans),
-            shape=self.shape,
-            dtype=self.dtype,
-            device=self.device,
-        )
-
-    def slice(self, start, end) -> "TensorRef":
-        # CONCAT path: Batch.slice → _slice_value → value.slice(start, end).
-        return self.select_ranges([(int(start), int(end))])
-
-    def __getitem__(self, key) -> "TensorRef":
-        # PACKED path: _slice_packed_data does ``value[cu[start]:cu[end]]``.
-        if isinstance(key, slice):
-            if key.step not in (None, 1):
-                raise NotImplementedError("TensorRef supports only contiguous (step=1) slicing")
-            start = 0 if key.start is None else int(key.start)
-            stop = self.batch_size if key.stop is None else int(key.stop)
-            return self.select_ranges([(start, stop)])
-        raise NotImplementedError(f"TensorRef indexing supports slices only, got {type(key).__name__}")
-
-    def transform(self, fn: Callable[[torch.Tensor], torch.Tensor]) -> "TensorRef":
-        backend = TensorTransportRuntime.current()
-        if backend is None:
-            raise RuntimeError("No TensorTransport backend installed")
-        return backend.transform(self, fn)
-
-    def reshape(self, *shape: int) -> "TensorRef":
-        return self.transform(lambda t: t.reshape(*shape))
-
-    def permute(self, *dims: int) -> "TensorRef":
-        return self.transform(lambda t: t.permute(*dims))
-
-    def local(self) -> torch.Tensor:
-        return self.materialize()
-
-    def materialize(self, backend: "Optional[TensorTransport]" = None) -> torch.Tensor:
-        """Fetch this meta into a real tensor (driver- or worker-side).
-
-        With a backend: one ``get`` over the spans (backends resolve a
-        :class:`TensorSpan` by resolving its handle and slicing). Without:
-        per-span ``local()`` round-trips assembled via :func:`cat_rows` (its
-        ragged right-pad contract applies to 2D+ per-shard-padded fields).
-        """
-        if backend is None:
-            backend = TensorTransportRuntime.current()
-        if not self.spans:
-            return torch.empty(0)
-        if backend is not None:
-            return backend.get(self.spans)
-        return cat_rows([s.local() for s in self.spans])
-
-    def retain_grad(self) -> "TensorRef":
-        self.retain_grad_flag = True
-        return self
-
-    @classmethod
-    def from_handles(cls, handles: list) -> "TensorRef":
-        """Wrap freshly-put handles as full-range spans — the single wrap chokepoint."""
-        spans = [TensorSpan(h, 0, int(h.shape[0])) for h in handles]
-        return cls(
-            spans=spans,
-            shape=(sum(int(h.shape[0]) for h in handles), *handles[0].shape[1:]) if handles else None,
-            dtype=handles[0].dtype if handles else None,
-            device=str(handles[0].device) if handles else None,
-        )
-
-    def __len__(self) -> int:
-        return self.batch_size
-
-
-# ---------------------------------------------------------------------------
-# Type-based tree walker
-# ---------------------------------------------------------------------------
-
 
 def _collect_leaves(
     value: Any,
@@ -456,15 +128,6 @@ def _collect_list(
 # ---------------------------------------------------------------------------
 
 
-def _apply_tensor_op(t: torch.Tensor, op: str, *args) -> torch.Tensor:
-    """Apply a named tensor op. Shared by the default ``tensor_op`` round-trip."""
-    if op == "getitem":
-        return t[args[0]]
-    if op == "reshape":
-        return t.reshape(args[0])
-    if op == "permute":
-        return t.permute(args[0])
-    raise ValueError(f"Unknown tensor op: {op!r}")
 
 
 class TensorTransport(abc.ABC):
@@ -645,226 +308,6 @@ class TensorTransport(abc.ABC):
             sess._flush()
 
 
-def map_tree(obj: Any, leaf_fn: Callable[[Any], Any]) -> Any:
-    """Rebuild a value tree, applying ``leaf_fn`` to every node.
-
-    The single tree-walker shared by the transport layer's rewrite passes
-    (controller-side ``localize`` substitution and ``Handle._rebind_tree``,
-    worker-side resolve/pack in ``Worker.call``). ``leaf_fn`` runs on every node
-    first; if it returns a *different* object that replaces the node and recursion
-    stops there. Otherwise containers are rebuilt structurally — ``Batch`` via
-    :meth:`Batch._rebuild` (preserving framework-managed ``_packed_cu_seqlens``),
-    ``tuple`` / ``list`` / ``dict`` element-wise. ``TensorRef`` is an atomic leaf
-    (never recursed into, despite being a ``Batch`` subclass); any other
-    non-container value passes through. Functional (returns new trees), so it works
-    on immutable tuples and lets each caller's ``leaf_fn`` decide what to swap.
-    """
-    new = leaf_fn(obj)
-    if new is not obj:
-        return new
-    if isinstance(obj, TensorRef):
-        return obj
-    if isinstance(obj, Batch):
-        return obj._rebuild({f.name: map_tree(getattr(obj, f.name), leaf_fn) for f in dc_fields(obj)})
-    if isinstance(obj, tuple):
-        return tuple(map_tree(item, leaf_fn) for item in obj)
-    if isinstance(obj, list):
-        return [map_tree(item, leaf_fn) for item in obj]
-    if isinstance(obj, dict):
-        return {k: map_tree(v, leaf_fn) for k, v in obj.items()}
-    return obj
-
-
-class WorkerLocalTransport(TensorTransport):
-    """The V2 Worker/Handle storage contract — worker-resident backends only.
-
-    Adds the storage-engine machinery only a worker-resident store needs:
-    ref-count lifecycle (``incref``/``decref``), controller-orchestrated
-    cross-worker transfer (``setup_transfer``/``nccl_send``/``nccl_recv``), and
-    on-worker remote compute (``tensor_op``/``cat``/``get_cpu``). The universal
-    materialization surface (``get_batch``/``put_batch``) lives on
-    the base. GLOBAL backends (e.g. the transfer queue) are plain
-    :class:`TensorTransport` and implement none of this capability.
-
-    ``isinstance(t, WorkerLocalTransport)`` is the locality discriminator the
-    controller uses to decide whether cross-worker routing is required.
-    """
-
-    # Methods the controller may invoke on this transport via the Worker actor's
-    # ``transport_op`` relay (TensorHandle GC/compute + Handle NCCL routing).
-    # Adding a capability method below means adding its name here. Excludes
-    # setup_transfer, which the Worker injects identity into via setup_global_pg.
-    REMOTE_OPS: ClassVar[frozenset] = frozenset({"incref", "decref", "tensor_op", "get_cpu", "nccl_send", "nccl_recv"})
-
-    # ---- lifecycle (ref-counting) ------------------------------------------
-
-    def incref(self, key: Any) -> None:
-        """Increment the ref count for a stored tensor. No-op by default."""
-
-    def decref(self, key: Any) -> None:
-        """Decrement the ref count; free at zero. No-op by default."""
-
-    # ---- locality + cross-worker transfer (localize) -------------------
-
-    def setup_transfer(self, global_rank: int, world_size: int) -> None:
-        """Initialize the cross-worker transfer group."""
-
-    def nccl_send(self, dst_rank: int, handles: List[Any]) -> None:
-        raise NotImplementedError("transport does not support cross-worker send")
-
-    def nccl_recv(self, src_rank: int, shapes: List[tuple], dtypes: List[torch.dtype]) -> List[Any]:
-        raise NotImplementedError("transport does not support cross-worker recv")
-
-    @classmethod
-    def _is_local(cls, ref: Any, dst_worker_id: str, dst_device_id: int, pool: Any) -> bool:
-        """True if ``ref`` is already resolvable on the dst worker (no transfer needed).
-
-        The one per-backend locality decision. Base: a ref is local only if produced by
-        the dst worker (per-process store). gpu overrides to also accept same physical
-        device, since its per-GPU TensorWorker is shared across that GPU's slots.
-        """
-        return ref.source_id == dst_worker_id
-
-    @classmethod
-    def _move_key(cls, span: Any, dst: Tuple[str, int], pool: Any) -> Optional[tuple]:
-        """Transfer identity of a span wrt a destination — or ``None`` if already resolvable there.
-
-        The single span classifier shared by ``localize``'s FIND and REPLACE passes. Pure
-        (no Ray): reads only the handle's identity fields, ``pool.device_id_of``, and the
-        per-backend ``_is_local``. Checked in short-circuit order so ``device_id_of`` is
-        never called on a resolvable ref:
-
-          1. ``object_ref`` present (CPU/plasma) → resolvable anywhere → ``None``
-          2. ``_is_local`` on the dst worker → ``None``
-          3. else → the value key ``(src_device, dst_device, store_key, start, stop)``
-
-        The key is by VALUE, not ``id()``: identical foreign slices headed to the same dst
-        device collapse to one transfer, and the received result is shared across every
-        tree position. ``store_key`` is unique per source device, and ``object_ref`` handles
-        (``store_key is None``) never reach branch 3, so the key cannot alias.
-        """
-        dst_worker_id, dst_device_id = dst
-        h = span.handle
-        if getattr(h, "object_ref", None) is not None:
-            return None
-        if cls._is_local(h, dst_worker_id, dst_device_id, pool):
-            return None
-        return (pool.device_id_of(h.source_id), dst_device_id, h.store_key, span.start, span.stop)
-
-    @classmethod
-    def _replace_leaf(cls, moved: Dict[tuple, Any], dst: Tuple[str, int], pool: Any) -> Callable[[Any], Any]:
-        """Build the ``map_tree`` leaf for REPLACE: swap each foreign span for its moved result.
-
-        A factory (not an inline loop closure) so the per-shard ``dst`` is a real parameter —
-        no loop late-binding. A ref whose spans all stay put is returned UNCHANGED (same
-        object): this skips a needless rebuild and preserves the fields ``with_spans`` drops
-        (``grad`` / ``retain_grad_flag`` / ``_packed_cu_seqlens``). Keys are recomputed on the
-        original spans, so they match exactly what FIND recorded.
-        """
-
-        def leaf(o: Any) -> Any:
-            if isinstance(o, TensorRef):
-                new_spans = [moved.get(cls._move_key(s, dst, pool), s) for s in o.spans]
-                if all(ns is s for ns, s in zip(new_spans, o.spans)):
-                    return o
-                return o.with_spans(new_spans)
-            return o
-
-        return leaf
-
-    @classmethod
-    def _move(cls, pool: Any, to_move: Dict[tuple, Any]) -> Dict[tuple, Any]:
-        """Run one batched NCCL hop per ``(src_device, dst_device)`` group; return key → received span.
-
-        Ordering invariant: each group's ``keys`` list is built once and reused — in the
-        SAME order — for (a) the ``nccl_send`` items, (b) the ``nccl_recv`` shapes/dtypes,
-        and (c) the ``zip(keys, recv_handles)`` that fills ``moved``. ``dict`` preserves
-        insertion order, which is what makes this sound; do not reorder one without the
-        others. All sends AND recvs are posted before any ``ray.get`` so a send cannot block
-        on an unposted recv.
-
-        Recv shapes/dtypes come off the representative span's ``.shape`` / ``.dtype`` — the
-        SLICED shape (exactly the rows the send ships), not the full handle block.
-        """
-        groups: Dict[Tuple[int, int], List[tuple]] = {}  # (src_dev, dst_dev) → [key, ...] (ordered)
-        for key in to_move:
-            groups.setdefault((key[0], key[1]), []).append(key)
-
-        send_refs, recv_refs = [], []
-        for (src_device_id, dst_device_id), keys in groups.items():
-            spans = [to_move[k] for k in keys]
-            send_refs.append(pool.slot0_worker(src_device_id).transport_op.remote("nccl_send", dst_device_id, spans))
-            recv_refs.append(
-                pool.slot0_worker(dst_device_id).transport_op.remote(
-                    "nccl_recv", src_device_id, [s.shape for s in spans], [s.dtype for s in spans]
-                )
-            )
-        ray.get(send_refs)
-        recv_results = ray.get(recv_refs)
-
-        moved: Dict[tuple, Any] = {}
-        for ((src_device_id, dst_device_id), keys), new_handles in zip(groups.items(), recv_results):
-            dst_worker = pool.slot0_worker(dst_device_id)
-            for key, new_h in zip(keys, new_handles):
-                new_h.rebind(dst_worker)
-                # The recv handle holds exactly the sliced rows → full-range span.
-                moved[key] = TensorSpan(new_h, 0, int(new_h.shape[0]))
-        return moved
-
-    @classmethod
-    def localize(cls, shards: list, pool: Any, device_ids: List[int], worker_ids: List[str]) -> list:
-        """Make every ref in each shard resolvable on its dst worker — find / move / replace.
-
-        Shared skeleton for all worker-local backends; the only thing that varies is
-        ``_is_local`` (the locality predicate, reached via ``_move_key``). Three stages:
-
-          FIND     read-only walk → the unique foreign slices to move, keyed by transfer
-                   identity; identical slices to one dst device dedup (``_move_key``).
-          MOVE     one batched NCCL hop per ``(src_device, dst_device)`` group between the
-                   two devices' slot0 workers (``_move``).
-          REPLACE  rebuild each shard, swapping foreign spans for their received result and
-                   leaving local spans (and untouched refs) as-is (``_replace_leaf``).
-
-        Names no backend type — works through ``span.handle`` / ``pool`` / ``map_tree`` /
-        the ``transport_op`` relay. FIND and REPLACE are pure (no Ray); only MOVE does I/O,
-        so when nothing is foreign the shards are returned untouched.
-        """
-        dsts = list(zip(worker_ids, device_ids))
-
-        # FIND — the unique foreign slices, keyed by transfer identity (setdefault dedups).
-        to_move: Dict[tuple, Any] = {}  # key → representative span
-        for (s_args, s_kwargs), dst in zip(shards, dsts):
-            for ref in collect_leaves(s_args, TensorRef) + collect_leaves(s_kwargs, TensorRef):
-                for s in ref.spans:
-                    key = cls._move_key(s, dst, pool)
-                    if key is not None:
-                        to_move.setdefault(key, s)
-        if not to_move:
-            return shards
-
-        # MOVE — one batched NCCL hop per (src_device, dst_device) group.
-        moved = cls._move(pool, to_move)
-
-        # REPLACE — rebuild each shard, swapping foreign spans for their received result.
-        return [
-            (map_tree(s_args, cls._replace_leaf(moved, dst, pool)), map_tree(s_kwargs, cls._replace_leaf(moved, dst, pool)))
-            for (s_args, s_kwargs), dst in zip(shards, dsts)
-        ]
-
-    # ---- remote compute (controller-triggered) ----------------------------
-
-    def tensor_op(self, handle: Any, op: str, *op_args) -> Any:
-        """Apply a named op (getitem/reshape/permute) to a stored tensor.
-
-        Default: round-trip resolve -> op -> put. Backends with on-worker compute
-        override to avoid moving data.
-        """
-        result = _apply_tensor_op(self._resolve_handles([handle])[0], op, *op_args).contiguous()
-        return self.put(result)
-
-    def get_cpu(self, handle: Any) -> torch.Tensor:
-        """Return the stored tensor as a CPU tensor."""
-        return self._resolve_handles([handle])[0].cpu()
 
 
 class TransportSession:
@@ -924,14 +367,5 @@ class TensorTransportRuntime:
         cls._current = None
 
 
-__all__ = [
-    "TensorHandle",
-    "TensorSpan",
-    "TensorRef",
-    "TensorTransport",
-    "TensorTransportRuntime",
-    "TransportSession",
-    "WorkerLocalTransport",
-    "cat_rows",
-    "map_tree",
-]
+
+__all__ = ["TensorTransport", "TensorTransportRuntime", "TransportSession"]
